# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Visit Lab",
        "contact": {
            "email": "saviogmoiagaia@gmail.com"
        },
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Loga um  usuário.",
                "description": "Loga",
                "operationId": "7716aff3e5de668956e4c3802154391b",
                "requestBody": {
                    "description": "Json com login e senha do usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Login do usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha do usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Formulário inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtém um user pelo id",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Um usuario"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Access token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "api_key": {
                "type": "apiKey",
                "description": "Chave de aplicação para autenticação na API",
                "name": "X-Application-Key",
                "in": "header"
            }
        }
    }
}